🚨 Lint files
